Refactor: The IDs are in the INT format.

+++ src/main/java/com/example/finanx/dto/WalletRecord.java
@@ -4,7 +4,7 @@ import java.util.List;
 import com.example.finanx.entities.Card;
 import com.example.finanx.entities.Wallet;
 
-public record WalletRecord(String owner, Double balance, List<Card> cards, Long userId){
+public record WalletRecord(String owner, Double balance, List<Card> cards, Integer userId){
   public WalletRecord(Wallet wallet) {
         this(wallet.getOwner(), wallet.getBalance(), wallet.getCards(), wallet.getUserId());
   }
